+++ bingads/v13/bulk/entities/ad_extensions/__init__.py
@@ -12,3 +12,4 @@ from .bulk_price_ad_extensions import *
 from .bulk_action_ad_extensions import *
 from .bulk_promotion_ad_extensions import *
 from .bulk_filterlink_ad_extensions import *
+from .bulk_flyer_ad_extensions import *